Step Definition segregated as Test Case belongs to  2 separate pages, and we will use variables in TestContextSetup class to traverse through the step definitions
